backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg[binary]==3.1.19
python-dotenv==1.0.0
pandas==2.1.3
python-multipart==0.0.6
jinja2==3.1.2
chardet==5.2.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1